# openapi.yaml
openapi: 3.0.3
info:
  title: Surge Loan Activity Log API
  version: 0.2.1
  description: |
    HTTP API backing agent activity logs and admin summaries.

    User onboarding exposes `/auth/register` with Argon2id-backed passwords and a short-lived JWT access token in JSON.
    The rotating refresh token is issued only as an HttpOnly `surge_refresh` cookie (Path `/auth`), not in response bodies, so browser JS cannot read it alongside the access token.
    Opaque refresh entries live in Redis. Roles (`user`, `admin`) attach at signup; elevate admins via DB or a future admin API.

    Sharing these docs publicly

    Deploy this service, then anyone with the HTTPS URL may use:

    * `/api-docs` — interactive Swagger UI (share link to this path).
    * `/openapi.json` — OpenAPI 3 JSON (Swagger Editor URL import, Postman, codegen).
    * `/openapi.yaml` — same spec as YAML.

    Both spec endpoints attach `Access-Control-Allow-Origin: *` so browsers on other sites may fetch the description (Swagger Editor, etc.) as long as your API URL is reachable. Update the `servers` entry below once you pick a canonical base URL (for display only—the server still serves whichever host resolves).

    Offline: paste or upload `openapi.yaml` from repo root into [Swagger Editor](https://editor.swagger.io/).
servers:
  - url: http://localhost:3000
    description: Local / development (replace after deployment)
paths:
  /openapi.yaml:
    get:
      tags: [Documentation]
      summary: OpenAPI 3 YAML document
      description: Public YAML copy suitable for SCM, Postman URL import where supported.
      responses:
        "200":
          description: OpenAPI YAML
          content:
            application/yaml:
              schema:
                type: string
  /openapi.json:
    get:
      tags: [Documentation]
      summary: OpenAPI 3 JSON document
      description: Public JSON preferred by many toolchain integrations.
      responses:
        "200":
          description: OpenAPI schema as JSON object
          content:
            application/json:
              schema:
                type: object
  /health:
    get:
      summary: Process health
      tags: [System]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
  /ready:
    get:
      summary: Dependency readiness (Postgres + Redis)
      tags: [System]
      responses:
        "200":
          description: Postgres and Redis reachable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ready
                  postgres:
                    type: boolean
                    example: true
                  redis:
                    type: boolean
                    example: true
        "500":
          description: Postgres or Redis unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      tags: [Auth]
      summary: Sign up — default role user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Account created — access JWT JSON plus HttpOnly refresh cookie surge_refresh (not in body)
          headers:
            Set-Cookie:
              description:
                surge_refresh opaque token HttpOnly SameSite=Lax Path /auth rotates on each login or refresh Secure in production
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokensResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Payload validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many signup attempts (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      summary: Email + password login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialsRequest"
      responses:
        "200":
          description: Access JWT JSON plus HttpOnly refresh cookie surge_refresh
          headers:
            Set-Cookie:
              description:
                surge_refresh rotates on login or refresh
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokensResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Payload validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many attempts (rate limited)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate refresh cookie and issue a fresh access JWT
      description:
        Requires the surge_refresh cookie from register or login. Send credentials (browser Cookie header). No JSON body.
      responses:
        "200":
          description: Access JWT refreshed; replaces surge_refresh cookie (rotating session)
          headers:
            Set-Cookie:
              description: Updated surge_refresh
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokensResponse"
        "401":
          description: Missing invalid or revoked refresh cookie
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke Redis refresh slot clear HttpOnly cookie
      description:
        Reads surge_refresh cookie if present clears cookie regardless. Idempotent.
      responses:
        "204":
          description: Session slot cleared cookie expired client-side
          headers:
            Set-Cookie:
              description: surge_refresh cleared
              schema:
                type: string
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user profile — requires Bearer access JWT
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Public profile snapshot (never includes password_hash)
          content:
            application/json:
              schema:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: "#/components/schemas/UserPublic"
        "401":
          description: Missing, malformed, or expired access JWT
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: JWT valid but backing user row disappeared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
tags:
  - name: Documentation
    description: Public API description files (Swagger / codegen).
  - name: System
    description: Operational endpoints
  - name: Auth
    description: Registration login JWT access plus HttpOnly refresh cookie rotation and logout
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description:
        Access JWT goes in Authorization Bearer. Obtain it via /auth/register, /auth/login,
        or /auth/refresh. Long-lived rotation uses the HttpOnly surge_refresh cookie (never returned in JSON).
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
    RegisterRequest:
      type: object
      required:
        - full_name
        - email
        - password
        - location_station
      properties:
        full_name:
          type: string
          maxLength: 200
          example: "Jamie Rivera"
        email:
          type: string
          format: email
          maxLength: 320
          example: "jamie@example.com"
        password:
          type: string
          minLength: 12
          maxLength: 200
          format: password
          description: Argon2id hash stored server-side; minimum 12 characters in v1
        location_station:
          type: string
          maxLength: 200
          example: "North Station Desk 4"
    CredentialsRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AccessTokensResponse:
      type: object
      description: Access JWT payload only refresh token is never included (HttpOnly cookie instead)
      required:
        - user
        - access_token
        - expires_in
      properties:
        user:
          $ref: "#/components/schemas/UserPublic"
        access_token:
          type: string
          description: HS256 JWT including roles — short-lived
        expires_in:
          type: integer
          description: Access TTL seconds
          example: 900
    UserPublic:
      type: object
      required:
        - id
        - email
        - full_name
        - location_station
        - roles
        - email_verified_at
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        location_station:
          type: string
        roles:
          type: array
          items:
            type: string
            example: user
          description: Role codes intersected by `requireRoles` middleware (`user`, seeded `admin` available)
        email_verified_at:
          type: string
          format: date-time
          nullable: true
          description: Set when SMTP verification arrives (null in onboarding v1)